requirements.txt
notion-client
supabase
python-dotenv
httpx
